Matrix 2: Created mMod_0_comp_TextEditorForMatrixModAmountSlider.h and .cpp and declared and defined class TextEditorForMatrixModAmountSlider. In SliderForMatrixModAmount, added an instance of TextEditorForMatrixModAmountSlider and declared and defined mouseDoubleClick().

+++ Matrix-6G/Matrix-6G 2.0/Source/splits/splits_0_build_ChoiceName.h
@@ -10,13 +10,13 @@ using namespace Matrix_6G_Constants;
 
 struct SplitParamChoiceName
 {
-	static String buildForZoneLimit(int choiceNum, bool verbose) {
+	static String buildForZoneLimit(uint8 choiceNum, bool verbose) {
 		jassert(choiceNum > -1 && choiceNum < SPLT::numberOfChoicesForZoneLimit);
 		auto pitchName{ ExposedParamChoiceNamesValueTree::convertIntToPitchName(choiceNum) };
 		return pitchName + (verbose ? " (MIDI Note " + (String)choiceNum + ")" : "");
 	}
 
-	static String buildForZoneTranspose(int choiceNum, bool verbose) {
+	static String buildForZoneTranspose(uint8 choiceNum, bool verbose) {
 		jassert(choiceNum > -1 && choiceNum < 61);
 		if (choiceNum < 35)
 			return (String)(choiceNum - 36) + (verbose ? " semitones" : "");
